requirements.txt
einops==0.6.1
matplotlib==3.7.1
numpy==1.24.3
opencv_python==4.7.0.72
Pillow==9.4.0
scipy==1.10.1
timm==0.9.2
torch==1.12.1
torchvision==0.13.1
tqdm==4.65.0